Move shader assets preparation code to assets.rs
